Pulled contributions validation into shared validation helper, and fixed bug with number input in IE (added isNaN check).
