+++ assets/components/contribAmounts/contribAmounts.jsx
@@ -71,7 +71,7 @@ const contribErrors = {
   tooLittleRecurring: 'Please enter at least £5',
   tooLittleOneOff: 'Please enter at least £1',
   tooMuch: 'We are presently only able to accept contributions of £2000 or less',
-  noEntry: 'Please enter a numeric amount',
+  invalidEntry: 'Please enter a numeric amount',
 };
