+++ assets/pages/bundles-landing/reducers/reducers.js
@@ -7,7 +7,7 @@ import { combineReducers } from 'redux';
 import { abTestReducer as abTests } from 'helpers/abtest';
 import type { Contrib, ContribError, Amounts } from 'helpers/contributions';
 
-import validateContribution from '../helpers/validation';
+import validateContribution from 'helpers/contributions';
 import type { Action } from '../actions/bundlesLandingActions';
